+++ transport_graph.go
@@ -627,9 +627,6 @@ func applyArgsFromStruct(r *graph.UnifiedResolver[any], argsType reflect.Type) (
 		if gqlType == nil {
 			gqlType = goTypeToGraphQL(f.Type)
 		}
-		if gqlType != nil {
-			gqlType = graph.MaskType(name, f.Type, gqlType)
-		}
 		if gqlType == nil {
 			continue
 		}
@@ -935,7 +932,7 @@ func buildInputObjectForType(t reflect.Type) graphql.Input {
 					if name == "" {
 						continue
 					}
-					sub := graph.MaskType(name, ef.Type, goTypeToGraphQL(ef.Type))
+					sub := goTypeToGraphQL(ef.Type)
 					if sub == nil {
 						continue
 					}
@@ -950,7 +947,7 @@ func buildInputObjectForType(t reflect.Type) graphql.Input {
 		if name == "" {
 			continue
 		}
-		sub := graph.MaskType(name, f.Type, goTypeToGraphQL(f.Type))
+		sub := goTypeToGraphQL(f.Type)
 		if sub == nil {
 			continue
 		}